requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.0
Cython==3.0.2
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
dnspython==2.4.2
easydict==1.10
email-validator==2.0.0.post2
eventlet==0.33.3
faiss-cpu==1.7.4
fastapi==0.103.2
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.43.0
frozenlist==1.4.0
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.8.4
greenlet==3.0.0
grpcio==1.59.0
h11==0.14.0
h5py==3.9.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
humanfriendly==10.0
idna==3.4
imageio==2.31.5
imutils==0.5.4
insightface==0.7.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lit==17.0.2
llvmlite==0.41.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
mxnet-cu110==1.9.1
networkx==3.1
numba==0.58.0
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.14.1
onnxruntime-gpu==1.15.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.7
packaging==23.2
Pillow==10.0.1
prettytable==3.9.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-multipart==0.0.6
python-socketio==5.9.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.3
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==2.0.1
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.8
websocket-server==0.6.4
websockets==11.0.3
Werkzeug==3.0.0
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
